+++ app/assets/javascripts/results_table/environment.js.coffee
@@ -1,25 +1,14 @@
 $ ->
   # 表示するデータを指定し初期化
-  dataTables.environment = $("#environment_results").DataTable(
-    ajax:
-      url: "/report_type/environments/search.json"
-      data: (d) ->
-        d.taxonomy = $("#_taxonomy_id").val()
-        d.environment = $("#_environment_id").val()
-        d.biological_process = $("#_biological_process_id").val()
-        d.molecular_function = $("#_molecular_function_id").val()
-        d.cellular_component = $("#_cellular_component_id").val()
-        d.phenotype = $("#_phenotype_id").val()
-        return
+  drawInfo.environment =
+    dataTable: $("#environment_results").DataTable(
+      ajax:
+        url: "/report_type/environments/search.json"
+      columns: [
+        {data: "environment_link"}
+      ]
+    )
+    downloadCSV: '/report_type/environments/search.csv'
 
-      error: ->
-        alert "failing query..."
-        return
-
-    columns: [
-      {data: "environment_link"}
-    ]
-    paginationSlider: null
-  )
   $("#environment_results").parent().find(".result-download-container").append "<a>Download CSV</a>"
   return
