+++ app/controllers/report_type/phenotypes_controller.rb
@@ -1,2 +1,2 @@
-class ReportType::PhenotypesController < ApplicationController
+class ReportType::PhenotypesController < ReportType::BaseController
 end
