+++ app/models/report_type/environment.rb
@@ -1,19 +1,7 @@
 module ReportType
   class Environment < Base
     class << self
-      def count(meo_id: '', tax_id: '', bp_id: '', mf_id: '', cc_id: '', mpo_id: '')
-        sparql  = environment_count_sparql(meo_id, tax_id, bp_id, mf_id, cc_id, mpo_id)
-        results = query(sparql)
-
-        results.first[:hits_count]
-      end
-
-      def search(meo_id: '', tax_id: '', bp_id: '', mf_id: '', cc_id: '', mpo_id: '', limit: 25, offset: 0)
-        sparql  = environment_search_sparql(meo_id, tax_id, bp_id, mf_id, cc_id, mpo_id, limit, offset)
-        results = query(sparql)
-
-        return [] if results.empty?
-
+      def addition_information(results)
         results.map do |result|
           Base::Environment.new(result[:meo_id], result[:meo_name])
         end
