+++ app/views/report_type/genes/search.csv.csvbuilder
@@ -14,7 +14,7 @@ csv << [
 
 limit = 100
 (0..@hits_count).step(limit) do |offset|
-  ReportType::Gene.search(@args.merge(limit: limit, offset:offset)).each do |p|
+  ReportType::Gene.search(@args.merge(limit: limit, offset: offset)).each do |p|
     csv << [
       p.name,
       p.genes.map(&:id).join("\n"),
