+++ app/views/report_type/organisms/search.json.jbuilder
@@ -1,8 +1,7 @@
-json.sEcho @sEcho
 json.iTotalRecords @total_count
 json.iTotalDisplayRecords @hits_count
 json.aaData do |json|
-  json.array!(@proteins) do |p|
+  json.array!(@results) do |p|
     json.organism_link     link_to(p.tax.name, organism_path(p.tax.id), target: '_blank')
     json.environment_links link_list(p.envs, 'environment')
     json.phenotype_links   link_list(p.phenotypes, 'phenotype')
