+++ app/views/report_type/proteins/search.csv.csvbuilder
@@ -20,7 +20,7 @@ csv << [
 
 limit = 100
 (0..@hits_count).step(limit) do |offset|
-  ReportType::Protein.search(@args.merge(limit: limit, offset:offset)).each do |p|
+  ReportType::Protein.search(@args.merge(limit: limit, offset: offset)).each do |p|
     csv << [
       p.name,
       p.genes.map(&:id).join("\n"),
